Integrate React into the application
Setup a basic integration of React into the Spring application. Notable
changes include the package.json and webpack.config.js files that
configure the NodeJS stuff, and a src/main/js directory for the React
code. Also note that you now have to run './mvnw package' to reflect
changes in the React code.