Refactor ContentCard into ContentInfoCard
ContentCard was really being used for displaying information about the
currently viewed content (in the ViewContentPage). As such, it was not a
general-purpose card for viewing content (like what might be used for
viewing search/category results). Thus, it has been refactored into a
specific ContentInfoCard, leaving the more general card to imlpemented
later.