Refactor how categories display content
Add a CategoryCard component that fetches its own data, and make
Category2 use it. This way, the category can start displaying data
without having to wait for all of the data to be ready.

Note that this refactor currently breaks regular search results
(they still display titles, but not thumbnails). This is because this
refactor includes a change to the /search endpoint where it no longer
returns sketch data (aside from sketch IDs).