Refactor comments
Optimize /comments/add endpoint (it was double saving before). Also
refactor the AddComment component into the Comments component. This new
component now displays a list of Comment IDs (to be turned into actual
comment data in a future update) in addition to the form for adding a
new comment. Also make the form for adding a new comment actually work
(it was hard coded before).