Add create content on current level button
Show a button to the right of the current content that allows you to
create content on the current level (i.e. add new child content to the
parent of the current content). The button is only shown when there is
not any "visible" content to the right of the current content, subject
to some additional rules: when the current content is private you can
always see the button; when it is public, you can only see the button if
the parent content was made contributable or if you are the author of
the current public content.